First cut at refactoring unittests into a top-level utility. Individual tests co-located with their DUT. No functional changes.
