+++ src/main/scala/groundtest/Comparator.scala
@@ -5,6 +5,7 @@ import uncore.tilelink._
 import uncore.constants._
 import junctions._
 import rocket._
+import util.Timer
 import scala.util.Random
 import cde.{Parameters, Field}
