+++ src/main/scala/junctions/util.scala
@@ -244,128 +244,3 @@ class DecoupledHelper(val rvs: Seq[Bool]) {
     (rvs.filter(_ ne exclude) ++ includes).reduce(_ && _)
   }
 }
-
-class MultiWidthFifo(inW: Int, outW: Int, n: Int) extends Module {
-  val io = new Bundle {
-    val in = Decoupled(Bits(width = inW)).flip
-    val out = Decoupled(Bits(width = outW))
-    val count = UInt(OUTPUT, log2Up(n + 1))
-  }
-
-  if (inW == outW) {
-    val q = Module(new Queue(Bits(width = inW), n))
-    q.io.enq <> io.in
-    io.out <> q.io.deq
-    io.count := q.io.count
-  } else if (inW > outW) {
-    val nBeats = inW / outW
-
-    require(inW % outW == 0, s"MultiWidthFifo: in: $inW not divisible by out: $outW")
-    require(n % nBeats == 0, s"Cannot store $n output words when output beats is $nBeats")
-
-    val wdata = Reg(Vec(n / nBeats, Bits(width = inW)))
-    val rdata = Vec(wdata.flatMap { indat =>
-      (0 until nBeats).map(i => indat(outW * (i + 1) - 1, outW * i)) })
-
-    val head = Reg(init = UInt(0, log2Up(n / nBeats)))
-    val tail = Reg(init = UInt(0, log2Up(n)))
-    val size = Reg(init = UInt(0, log2Up(n + 1)))
-
-    when (io.in.fire()) {
-      wdata(head) := io.in.bits
-      head := head + UInt(1)
-    }
-
-    when (io.out.fire()) { tail := tail + UInt(1) }
-
-    size := MuxCase(size, Seq(
-      (io.in.fire() && io.out.fire()) -> (size + UInt(nBeats - 1)),
-      io.in.fire() -> (size + UInt(nBeats)),
-      io.out.fire() -> (size - UInt(1))))
-
-    io.out.valid := size > UInt(0)
-    io.out.bits := rdata(tail)
-    io.in.ready := size < UInt(n)
-    io.count := size
-  } else {
-    val nBeats = outW / inW
-
-    require(outW % inW == 0, s"MultiWidthFifo: out: $outW not divisible by in: $inW")
-
-    val wdata = Reg(Vec(n * nBeats, Bits(width = inW)))
-    val rdata = Vec.tabulate(n) { i =>
-      Cat(wdata.slice(i * nBeats, (i + 1) * nBeats).reverse)}
-
-    val head = Reg(init = UInt(0, log2Up(n * nBeats)))
-    val tail = Reg(init = UInt(0, log2Up(n)))
-    val size = Reg(init = UInt(0, log2Up(n * nBeats + 1)))
-
-    when (io.in.fire()) {
-      wdata(head) := io.in.bits
-      head := head + UInt(1)
-    }
-
-    when (io.out.fire()) { tail := tail + UInt(1) }
-
-    size := MuxCase(size, Seq(
-      (io.in.fire() && io.out.fire()) -> (size - UInt(nBeats - 1)),
-      io.in.fire() -> (size + UInt(1)),
-      io.out.fire() -> (size - UInt(nBeats))))
-
-    io.count := size >> UInt(log2Up(nBeats))
-    io.out.valid := io.count > UInt(0)
-    io.out.bits := rdata(tail)
-    io.in.ready := size < UInt(n * nBeats)
-  }
-}
-
-// ============
-// Static timer
-// ============
-
-// Timer with a statically-specified period.
-// Can take multiple inflight start-stop events with ID
-// Will continue to count down as long as at least one event is inflight
-
-class Timer(initCount: Int, maxInflight: Int) extends Module {
-  val io = new Bundle {
-    val start = Valid(UInt(width = log2Up(maxInflight))).flip
-    val stop = Valid(UInt(width = log2Up(maxInflight))).flip
-    val timeout = Valid(UInt(width = log2Up(maxInflight)))
-  }
-
-  val inflight = Reg(init = Vec.fill(maxInflight) { Bool(false) })
-  val countdown = Reg(UInt(width = log2Up(initCount)))
-  val active = inflight.reduce(_ || _)
-
-  when (active) {
-    countdown := countdown - UInt(1)
-  }
-
-  when (io.start.valid) {
-    inflight(io.start.bits) := Bool(true)
-    countdown := UInt(initCount - 1)
-  }
-  when (io.stop.valid) {
-    inflight(io.stop.bits) := Bool(false)
-  }
-
-  io.timeout.valid := countdown === UInt(0) && active
-  io.timeout.bits := PriorityEncoder(inflight)
-
-  assert(!io.stop.valid || inflight(io.stop.bits),
-         "Timer stop for transaction that's not inflight")
-}
-
-object Timer {
-  def apply(initCount: Int, start: Bool, stop: Bool): Bool = {
-    val timer = Module(new Timer(initCount, 1))
-    timer.io.start.valid := start
-    timer.io.start.bits  := UInt(0)
-    timer.io.stop.valid  := stop
-    timer.io.stop.bits   := UInt(0)
-    timer.io.timeout.valid
-  }
-}
-
-
