+++ src/main/scala/uncore/tilelink/Drivers.scala
@@ -1,8 +1,7 @@
-package uncore.unittests
+package uncore.tilelink
 
 import Chisel._
 import junctions._
-import uncore.tilelink._
 import uncore.constants._
 import uncore.util._
 import cde.Parameters
